requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.12.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
razorpay==1.3.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0

